fix: unify service, fix golden test
Fixing a broken golden test, which should help ensure
that aepc continues to work as intended.

also unifying the example service for convenience, adding
diagrams to explain how it is architected.
