+++ DEVELOPMENT.md
@@ -20,6 +20,5 @@ go install github.com/bufbuild/buf/cmd/buf@v1.27.1
 
 The standard GoLang toolchain is used, with the addition of protobuf for
 compiling the resource definition.
-
-1. `protoc ./schema/resourcedefinition.proto --go_opt paths=source_relative --go_out=.`
+1. `./scripts/regenerate-all.sh`
 2. `go build main.go`
